move resource to lib
