+++ lib/csharp/microsoft.bot.builder.skills/Microsoft.Bot.Builder.Skills/Contextual/Dialogs/ResolveContextualInfoResponses.json
