+++ lib/csharp/microsoft.bot.builder.skills/Microsoft.Bot.Builder.Skills/Contextual/Dialogs/ResolveContextualInfoResponses.tt
@@ -1,3 +1,3 @@
 ﻿<#@ template debug="false" hostspecific="true" language="C#" #>
 <#@ output extension=".cs" #>
-<#@ include file="..\Shared\ResponseIdCollection.t4"#>
+<#@ include file=".\ResponseIdCollection.t4"#>
