+++ lib/csharp/microsoft.bot.builder.skills/Microsoft.Bot.Builder.Skills/Contextual/Dialogs/ResponseIdCollection.t4
@@ -1,0 +1,31 @@
+﻿<#@ assembly name="Newtonsoft.Json.dll" #>
+<#
+  var className = System.IO.Path.GetFileNameWithoutExtension(Host.TemplateFile);
+  var namespaceName = System.Runtime.Remoting.Messaging.CallContext.LogicalGetData("NamespaceHint");
+  string myFile = System.IO.File.ReadAllText(this.Host.ResolvePath(className + ".json"));
+  var json = Newtonsoft.Json.JsonConvert.DeserializeObject<System.Collections.Generic.Dictionary<string, dynamic>>(myFile);
+  var responses = string.Empty;
+  var cards = string.Empty;
+#>
+// https://docs.microsoft.com/en-us/visualstudio/modeling/t4-include-directive?view=vs-2017
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Microsoft.Bot.Builder.Solutions.Responses;
+
+namespace <#= namespaceName #>
+{
+    /// <summary>
+    /// Contains bot responses.
+    /// </summary>
+    public class <#= className #> : IResponseIdCollection
+    {
+        // Generated accessors
+<#
+// This code runs in the text json:  
+foreach (var propertyName in json)  { 
+#>
+        public const string <#= propertyName.Key.Substring(0, 1).ToUpperInvariant() + propertyName.Key.Substring(1) #> = "<#= propertyName.Key #>";
+<# } #>
+    }
+}
